net模块-协议管理类&协议模块集成测试

+++ net/src/main/java/com/awake/net/packet/common/Heartbeat.java
@@ -18,7 +18,7 @@ public class Heartbeat implements IPacket {
     public static final short PROTOCOL_ID = 102;
 
     @Override
-    public short protocolId() {
+    public int protocolId() {
         return PROTOCOL_ID;
     }
